Started working on Add Datapoint functionality

+++ src-db/src/models/data_points.rs
@@ -6,7 +6,7 @@ pub struct Model {
     #[sea_orm(primary_key)]
     pub id: u64,
     pub project: u16,
-    pub data: f64,
+    pub data: f32,
     pub created_at: chrono::DateTime<Utc>,
     pub updated_at: chrono::DateTime<Utc>,
 }
