+++ src-shared/src/data_point.rs
@@ -20,7 +20,7 @@ pub enum DataPointError {
 pub struct DataPoint {
     pub id: u64,
     pub project: u16,
-    pub data: f64,
+    pub data: f32,
     pub created_at: chrono::DateTime<Utc>,
     pub updated_at: chrono::DateTime<Utc>,
 }
